Changed how the data was manipulated.
The data that was created wasn't in the expected format for the YOLO model to read. Instead of it being encoded in the tensorflow native way, it was encoded through PIL which led to the model not understanding and inaccurate guesses. The data now uses the built in tensorflow method and this led to a significant increase in model performance.